Support multiple reporting CIK
The current logic was single honed to one of Vanguard's
reporting companies. To support more indexes, we need
to repeating the logic for each reporting company's CIK.

As EDGAR's URLs are inconsistent with whether to pad the
CIK up to 10 digits, we manipulate them as int and pad
them as needed.

While adding this logic mapped more ETFs and cleaned
up the script's standard output.