Add a cap on how many submisisions we can fetch
Despite fetching at a lower than required rate,
the script keeps hitting an undocumented cap on the
number of submissions we can fetch in a short window.

To make sure we fetch all indexes, the limit is a hard
one but we will fetch less than the limit as long as
we fetch all submissions on a fetch date.