Incrementally build history
The current logic couldn't handle building the history over time.

The way to handle building incrementally is:
1. The fetched map that contains the fetched interval of filings
2. Loading existing JSON so we can merge them together.

This requires rethinking how the code works and introduced a lot
of complexity. This will be likely handled in future PRs as
our current testing is sub-par.

This also underscored that we didn't sort our data from newest
to oldest, which impacted the fetched data.