Rework the global throttling
The global throttling logic wasn't well thought out
and could result in the IP being banned for 10 mins.
This was because the limits were per company rather
than global (!!!).

The solution is to move to actually do global
throtting and move it into edgar_client.

The benefit of this approach is that we shouldn't
run afoul of any limits. A side-change is that we
can potentially have a long running program as
we will do a long sleep to avoid the limit.