Add a throttler abstraction
This allows us to simplify the logic in the edgar
client and test the throttling logic at the same time
as we had a bunch of issues related to the throttler.

This PR adds a dependency on a fakable clock, which is
used by the throttler to ensure consistent behavior
between the normal usage and tests.